feat(crm): complete quote versioning and price snapshots

+++ Server/mitedtsm-module-crm/src/main/java/com/meession/etm/module/crm/controller/admin/contract/vo/contract/CrmContractSaveReqVO.java
@@ -107,7 +107,7 @@ public class CrmContractSaveReqVO {
 
         @Schema(description = "产品数量", requiredMode = Schema.RequiredMode.REQUIRED, example = "8911")
         @NotNull(message = "产品数量不能为空")
-        private Integer count;
+        private BigDecimal count;
 
     }
